Refactoring cozy init :dizzy:
