+++ mobile/src/actions/index.js
@@ -1,6 +1,7 @@
 /* global __ALLOW_HTTP__ */
 
-import cozy, { LocalStorage as Storage } from 'cozy-client-js'
+import cozy from 'cozy-client-js'
+import { init } from '../lib/cozy_init'
 import localforage from 'localforage'
 
 import { DISPLAY_TOAST } from '../../../src/actions'
@@ -47,56 +48,7 @@ export const setUrl = (url) => {
 export const registerDevice = (router, location) => {
   return async (dispatch, getState) => {
     await dispatch(setUrl(getState().mobile.serverUrl))
-
-    let oauth
-    if (window.cordova && window.cordova.InAppBrowser) {
-      oauth = {
-        storage: new Storage(),
-        clientParams: {
-          redirectURI: 'http://localhost',
-          softwareID: 'io.cozy.mobile.files',
-          clientName: 'Mobile Cozy Files',
-          scopes: ['io.cozy.files:GET']
-        },
-        onRegistered: (client, url) => {
-          if (window.cordova && window.cordova.InAppBrowser) {
-            const { InAppBrowser } = window.cordova
-            const target = '_blank'
-            const options = 'location=yes,hidden=no'
-            const inAppBrowser = InAppBrowser.open(url, target, options)
-
-            return new Promise((resolve) => {
-              inAppBrowser.addEventListener('loadstart', ({url}) => {
-                const accessCode = /\?access_code=(.+)$/.test(url)
-                const state = /\?state=(.+)$/.test(url)
-
-                if (accessCode || state) {
-                  resolve(url)
-                }
-              })
-            })
-            .then(
-              (url) => {
-                inAppBrowser.close()
-                return url
-              },
-              (err) => {
-                inAppBrowser.close()
-                dispatch(wrongAddressError())
-                throw err
-              }
-            )
-          }
-        }
-      }
-    }
-
-    cozy.init({
-      cozyURL: `${getState().mobile.serverUrl}`,
-      offline: {doctypes: ['io.cozy.files'], timer: 15},
-      oauth: oauth
-    })
-
+    await init(getState().mobile.serverUrl)
     try {
       await cozy.authorize()
       await cozy.offline.replicateFromCozy('io.cozy.files')
