+++ package.json
@@ -90,7 +90,7 @@
   "dependencies": {
     "classnames": "^2.2.0",
     "cozy-bar": "3.0.0-beta3",
-    "cozy-client-js": "0.0.7",
+    "cozy-client-js": "0.0.9",
     "cozy-ui": "3.0.0-beta6",
     "date-fns": "^1.22.0",
     "filesize": "^3.3.0",
